Work around stupid bug in QtConcurrent.
You have to call setProgressValue(AndText) at least once
unconditionally, before resultReporting and isProgressUpdateNeeded
actually do the right thing.

Also use qreal progress values internally while iterating directories.

+++ src/libs/utils/filesearch.h
@@ -72,9 +72,9 @@ public:
 private:
     QStringList m_filters;
     mutable QStack<QDir> m_dirs;
-    mutable QStack<int> m_progressValues;
+    mutable QStack<qreal> m_progressValues;
     mutable QStack<bool> m_processedValues;
-    mutable int m_progress;
+    mutable qreal m_progress;
     mutable QStringList m_currentFiles;
 };
